Optimize some singleton class.
