+++ app/src/main/java/com/symeonchen/wakeupscreen/ScBaseFragment.kt
@@ -3,7 +3,6 @@ package com.symeonchen.wakeupscreen
 import androidx.fragment.app.Fragment
 
 open class ScBaseFragment : Fragment() {
-
     companion object {
         var TAG: String = this::class.java.simpleName
     }
